Another Windows fixes
UNIQUE_NAME is defined somewhere in MinGW headers.

+++ src/video_capture/testcard.cpp
@@ -436,7 +436,7 @@ static int vidcap_testcard_init(struct vidcap_params *params, void **state)
                         fseek(in, 0L, SEEK_SET);
 
                         if (s->frame->tiles[0].data_len != static_cast<unsigned>(filesize)) {
-                                int level = s->frame->tiles[0].data_len < filesize ? LOG_LEVEL_WARNING : LOG_LEVEL_ERROR;
+                                int level = s->frame->tiles[0].data_len < static_cast<unsigned>(filesize) ? LOG_LEVEL_WARNING : LOG_LEVEL_ERROR;
                                 LOG(level) << MOD_NAME  << "Wrong file size for selected "
                                         "resolution and codec. File size " << filesize << ", "
                                         "computed size " << s->frame->tiles[0].data_len << "\n";
